+++ src/commands/database-settings-set.ts
@@ -3,10 +3,10 @@ import * as _ from "lodash";
 import * as request from "request";
 
 import * as responseToError from "../responseToError";
-import * as Command from "../command";
+import { Command } from "../command";
 import * as logger from "../logger";
 import { FirebaseError } from "../error";
-import * as requirePermissions from "../requirePermissions";
+import { requirePermissions } from "../requirePermissions";
 import * as utils from "../utils";
 import * as api from "../api";
 import * as requireInstance from "../requireInstance";