+++ src/commands/ext-uninstall.ts
@@ -1,14 +1,14 @@
 import * as _ from "lodash";
 import * as clc from "cli-color";
 import * as ora from "ora";
-import * as Command from "../command";
+import { Command } from "../command";
 import { FirebaseError } from "../error";
 import * as getProjectId from "../getProjectId";
 import { iam } from "../gcp";
 import * as extensionsApi from "../extensions/extensionsApi";
 import { ensureExtensionsApiEnabled, logPrefix } from "../extensions/extensionsHelper";
 import { promptOnce } from "../prompt";
-import * as requirePermissions from "../requirePermissions";
+import { requirePermissions } from "../requirePermissions";
 import * as utils from "../utils";
 
 export default new Command("ext:uninstall <extensionInstanceId>")