+++ src/commands/firestore-delete.js
@@ -1,10 +1,10 @@
 "use strict";
 
 var clc = require("cli-color");
-var Command = require("../command");
+var { Command } = require("../command");
 var FirestoreDelete = require("../firestore/delete");
 var { prompt } = require("../prompt");
-var requirePermissions = require("../requirePermissions");
+var { requirePermissions } = require("../requirePermissions");
 var utils = require("../utils");
 
 var _getConfirmationMessage = function(deleteOp, options) {