+++ src/commands/functions-config-unset.js
@@ -3,11 +3,11 @@
 var _ = require("lodash");
 
 var clc = require("cli-color");
-var Command = require("../command");
+var { Command } = require("../command");
 var functionsConfig = require("../functionsConfig");
 var getProjectId = require("../getProjectId");
 var logger = require("../logger");
-var requirePermissions = require("../requirePermissions");
+var { requirePermissions } = require("../requirePermissions");
 var utils = require("../utils");
 var runtimeconfig = require("../gcp/runtimeconfig");