+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Fixes a bug where the Functions emulator ignored the "host" configuration (#1722)
 * Fixes a bug where the Functions emulator accepted requests to too many paths (#1773)
+* Print Firebase Console links for Extensions after operations.
 * Adds the `firebase init emulators` command.
 * Adds a Cloud Pub/Sub Emulator (#1748).